% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMpalette.R
\name{colorblind.friendly}
\alias{colorblind.friendly}
\title{Colorblind-Friendly Palette Check}
\usage{
colorblind.friendly(palette_name)
}
\arguments{
\item{palette_name}{Name of Palette. Choices are:
\code{all_cols}, \code{primary_cols}, and \code{accent_cols},}
}
\value{
TRUE/FALSE value whether palette is colorblind-friendly
}
\description{
Checks whether a palette is colorblind-friendly. Colorblind-friendliness tested using the 'colorblindcheck' package.
To be colorblind-friendly, all colors in the palettes must be distinguishable with deuteranopia, protanopia, and tritanopia.
}
\examples{
colorblind.friendly("all_cols")
}
