% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMpalette.R
\name{em.brewer}
\alias{em.brewer}
\title{EM Palette Generator}
\usage{
em.brewer(
  palette_name,
  n,
  type = c("discrete", "continuous"),
  direction = c(1, -1),
  override.order = FALSE
)
}
\arguments{
\item{n}{Number of desired colors. If number of requested colors is beyond the scope of the palette,
colors are automatically interpolated. If n is not provided, the length of the palette is used.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want to automatically
interpolate between colors.}

\item{direction}{Sets order of colors. Default palette is 1. If direction is -1, palette color order is reversed}

\item{override.order}{Colors are picked from palette to maximize readability and aesthetics. This means
that colors are not always selected in sequential order from the full palette. If override.order is set to TRUE,
colors are selected in sequential order from the full palette instead. Default is FALSE.}

\item{palette}{Name of Palette. Choices are:
\code{all_cols}, \code{primary_cols}, and \code{accent_cols}.}
}
\value{
A vector of colors.
}
\description{
Color palettes inspired by Color pallets of ExxonMobil. This is an unofficial package based on the
branding guidelines found at \href{www.exxonmobil.com}{on Github}.
Use \code{\link{colorblind.friendly}} to check whether palettes are colorblind-friendly.
}
\examples{
em.brewer("all_cols")

em.brewer("accent_cols", direction=-1)

em.brewer("primary_cols", 4, override.order=TRUE)

library(ggplot2)
ggplot(data=iris, aes(x=Species, y=Petal.Length, fill=Species)) +
geom_violin() +
scale_fill_manual(values=em.brewer("primary_cols", 3))

ggplot(data=iris, aes(x=Sepal.Length, y=Sepal.Width, color=Species)) +
geom_point(size=2) +
scale_color_manual(values=em.brewer("accent_cols", 3))

ggplot(data=iris, aes(x=Species, y=Sepal.Width, color=Sepal.Width)) +
geom_point(size=3) +
scale_color_gradientn(colors=em.brewer("accent_cols"))
}
\keyword{colors}
