% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMpalette.R
\docType{data}
\name{EMpalettes}
\alias{EMpalettes}
\title{Complete list of palettes.}
\format{
An object of class \code{list} of length 3.
}
\usage{
EMpalettes
}
\description{
Use names(EMPalettes) to return all possible palette names. Current choices are:
\code{all_cols}, \code{primar_cols}, and \code{accent_cols}.
Use \code{\link{em.brewer}} to construct palettes.
}
\keyword{datasets}
