% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMpalette.R
\name{scale_color_em_c}
\alias{scale_color_em_c}
\title{EMpalette palettes for plotting with ggplot2}
\usage{
scale_color_em_c(palette_name, direction = 1, ...)
}
\arguments{
\item{palette_name}{Name of Palette. Choices are:
\code{all_cols}, \code{primary_cols}, and \code{accent_cols}.}

\item{direction}{Sets order of colors. Default palette is 1. If direction is -1, palette color order is reversed}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{scale_color_gradientn}}}
}
\description{
Function for using \code{EMPalettes} colors schemes in \code{ggplot2}. Use \code{\link{scale_color_em_d}} and \code{\link{scale_fill_em_d}}
for discrete scales and \code{\link{scale_color_em_c}} and \code{\link{scale_fill_em_c}} for continuous scales.
}
\examples{
library(ggplot2)
ggplot(data=iris, aes(x=Sepal.Length, y=Sepal.Width, color=Sepal.Length)) +
geom_point() +
scale_color_em_c("primary_cols", direction=-1)
}
