% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMpalette.R
\name{scale_fill_em_d}
\alias{scale_fill_em_d}
\title{EMBrewer palettes for plotting with ggplot2}
\usage{
scale_fill_em_d(palette_name, direction = 1, override.order = FALSE, ...)
}
\arguments{
\item{palette_name}{Name of Palette. Choices are:
\code{all_cols}, \code{primar_cols}, and \code{accent_cols}.}

\item{direction}{Sets order of colors. Default palette is 1. If direction is -1, palette color order is reversed}

\item{override.order}{Colors are picked from palette to maximize readability and aesthetics. This means
that colors are not always selected in sequential order from the full palette. If override.order is set to TRUE,
colors are selected in sequential order from the full palette instead. Default is FALSE.}

\item{...}{Other arguments passed on to \code{\link[ggplot2]{discrete_scale}}}
}
\description{
Function for using \code{EMPalette} colors schemes in \code{ggplot2}. Use \code{\link{scale_color_em_d}} and \code{\link{scale_fill_em_d}}
for discrete scales and \code{\link{scale_color_em_c}} and \code{\link{scale_fill_em_c}} for continuous scales.
}
\examples{
library(ggplot2)
ggplot(data=iris, aes(x=Species, y=Sepal.Length, fill=Species)) +
geom_violin() +
scale_fill_em_d("Lakota")
}
